Add csv version of participant declarations endpoint
