+++ app/controllers/api/v1/participants_controller.rb
@@ -7,6 +7,7 @@ module Api
     class ParticipantsController < Api::ApiController
       include ApiTokenAuthenticatable
       include ApiPagination
+      include ApiCsv
 
       def index
         respond_to do |format|
@@ -37,21 +38,6 @@ module Api
 
       def access_scope
         LeadProviderApiToken.joins(cpd_lead_provider: [:lead_provider])
-      end
-
-      def to_csv(hash)
-        return "" if hash[:data].empty?
-
-        headers = %w[id]
-        attributes = hash[:data].first[:attributes].keys
-        headers.concat(attributes.map(&:to_s))
-        CSV.generate(headers: headers, write_headers: true) do |csv|
-          hash[:data].each do |item|
-            row = [item[:id]]
-            row.concat(attributes.map { |attribute| item[:attributes][attribute].to_s })
-            csv << row
-          end
-        end
       end
 
       def updated_since
